move dhcp lease client code to a separate file
